Terminate driver on Ctrl-C.
With Perl scripts using system(), a Ctrl-C is sent only to the
child process, and the Perl script continues to run.  To completely
terminate the SHOC driver script required pressing Ctrl-C many times
to kill all the sub-processes for the benchmark problems.  Now
Ctrl-C terminates the benchmark program and the driver script.