Changed memory check to work with Jetson TK1 (issue #38)
There is an apparent cuda driver bug on the Jetson TK1 that results in
cudaMalloc() being handled incorrectly.  If more than the available
memory is requested, the process is killed instead of cudaMalloc
returning an error code as it is supposed to do.  This impedes the
function of findAvailBytes in cuda/common/support.h.